Add files via upload
Linear Regression is a machine learning technique used to model the relationship between two variables. It requires labeled data and aims to find a linear equation that best describes the relationship between the predictor and the response variable. It is widely used in many fields, particularly to predict a continuous variable based on one or more predictor variables.